refactor(web): decompose i18n module into smaller focused files
Split the monolithic i18n.ts (~1700 lines) into separate modules:
- types.ts: type definitions
- languages.ts: language options and helpers
- translate.ts: locale get/set and translation logic
- locales/: one file per locale
- index.ts: re-exports for backward-compatible imports

Update LanguageSelector to use option.label directly and remove
unused getLanguageOptionLabel import and id attribute.
