+++ web/src/lib/i18n/index.ts
@@ -1,0 +1,4 @@
+export type { Locale, LocaleDirection, LanguageOption, LocaleDocumentTarget } from './types';
+export { LANGUAGE_OPTIONS, LANGUAGE_SWITCH_ORDER, getLocaleDirection, getLanguageOption } from './languages';
+export { coerceLocale, getLocale, setLocale, t, tLocale, applyLocaleToDocument } from './translate';
+export { translations } from './locales';
