+++ web/src/lib/i18n/locales/ar.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const ar: Partial<TranslationKeys> = {
+  'nav.dashboard': 'لوحة التحكم',
+  'nav.agent': 'الوكيل',
+  'nav.tools': 'الأدوات',
+  'nav.cron': 'المهام المجدولة',
+  'nav.integrations': 'التكاملات',
+  'nav.memory': 'الذاكرة',
+  'nav.config': 'الإعدادات',
+  'nav.cost': 'تتبع التكلفة',
+  'nav.logs': 'السجلات',
+  'nav.doctor': 'التشخيص',
+  'dashboard.hero_title': 'لوحة تشغيل كهربائية',
+  'agent.placeholder': 'اكتب رسالة…',
+  'tools.search': 'ابحث في الأدوات…',
+  'cron.add': 'إضافة مهمة',
+  'memory.add_memory': 'إضافة ذاكرة',
+  'config.save': 'حفظ',
+  'cost.token_statistics': 'إحصاءات الرموز',
+  'logs.title': 'السجلات المباشرة',
+  'doctor.title': 'تشخيص النظام',
+  'auth.pair_button': 'اقتران',
+  'auth.enter_code': 'أدخل رمز الاقتران لمرة واحدة من الطرفية',
+  'auth.code_placeholder': 'رمز من 6 أرقام',
+  'auth.pairing_progress': 'جارٍ الاقتران…',
+  'auth.logout': 'تسجيل الخروج',
+  'common.languages': 'اللغات',
+  'common.select_language': 'اختر اللغة',
+  'header.dashboard_tagline': 'لوحة ZeroClaw',
+};
+
+export default ar;
