+++ web/src/lib/i18n/locales/bn.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const bn: Partial<TranslationKeys> = {
+  'nav.dashboard': 'ড্যাশবোর্ড',
+  'nav.agent': 'এজেন্ট',
+  'nav.tools': 'টুলস',
+  'nav.cron': 'নির্ধারিত কাজ',
+  'nav.integrations': 'ইন্টিগ্রেশন',
+  'nav.memory': 'মেমরি',
+  'nav.config': 'কনফিগারেশন',
+  'nav.cost': 'খরচ ট্র্যাকার',
+  'nav.logs': 'লগ',
+  'nav.doctor': 'ডায়াগনস্টিক',
+  'dashboard.hero_title': 'ইলেকট্রিক রানটাইম ড্যাশবোর্ড',
+  'agent.placeholder': 'একটি বার্তা লিখুন…',
+  'tools.search': 'টুল খুঁজুন…',
+  'cron.add': 'কাজ যোগ করুন',
+  'memory.add_memory': 'মেমরি যোগ করুন',
+  'config.save': 'সংরক্ষণ করুন',
+  'cost.token_statistics': 'টোকেন পরিসংখ্যান',
+  'logs.title': 'লাইভ লগ',
+  'doctor.title': 'সিস্টেম ডায়াগনস্টিক',
+  'auth.pair_button': 'পেয়ার করুন',
+  'auth.enter_code': 'টার্মিনাল থেকে একবারের পেয়ারিং কোড লিখুন',
+  'auth.code_placeholder': '৬-সংখ্যার কোড',
+  'auth.pairing_progress': 'পেয়ার করা হচ্ছে…',
+  'auth.logout': 'লগ আউট',
+  'common.languages': 'ভাষাসমূহ',
+  'common.select_language': 'ভাষা বেছে নিন',
+  'header.dashboard_tagline': 'ZeroClaw ড্যাশবোর্ড',
+};
+
+export default bn;
