+++ web/src/lib/i18n/locales/cs.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const cs: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Nástěnka',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Nástroje',
+  'nav.cron': 'Plánované úlohy',
+  'nav.integrations': 'Integrace',
+  'nav.memory': 'Paměť',
+  'nav.config': 'Konfigurace',
+  'nav.cost': 'Náklady',
+  'nav.logs': 'Logy',
+  'nav.doctor': 'Diagnostika',
+  'dashboard.hero_title': 'Elektrický runtime panel',
+  'agent.placeholder': 'Napište zprávu…',
+  'tools.search': 'Hledat nástroje…',
+  'cron.add': 'Přidat úlohu',
+  'memory.add_memory': 'Přidat paměť',
+  'config.save': 'Uložit',
+  'cost.token_statistics': 'Statistiky tokenů',
+  'logs.title': 'Živé logy',
+  'doctor.title': 'Diagnostika systému',
+  'auth.pair_button': 'Spárovat',
+  'auth.enter_code': 'Zadejte jednorázový párovací kód z terminálu',
+  'auth.code_placeholder': '6místný kód',
+  'auth.pairing_progress': 'Párování…',
+  'auth.logout': 'Odhlásit se',
+  'common.languages': 'Jazyky',
+  'common.select_language': 'Vyberte jazyk',
+  'header.dashboard_tagline': 'Panel ZeroClaw',
+};
+
+export default cs;
