+++ web/src/lib/i18n/locales/da.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const da: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Kontrolpanel',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Værktøjer',
+  'nav.cron': 'Planlagte job',
+  'nav.integrations': 'Integrationer',
+  'nav.memory': 'Hukommelse',
+  'nav.config': 'Konfiguration',
+  'nav.cost': 'Omkostninger',
+  'nav.logs': 'Logge',
+  'nav.doctor': 'Diagnostik',
+  'dashboard.hero_title': 'Elektrisk runtime-kontrolpanel',
+  'agent.placeholder': 'Skriv en besked…',
+  'tools.search': 'Søg værktøjer…',
+  'cron.add': 'Tilføj job',
+  'memory.add_memory': 'Tilføj hukommelse',
+  'config.save': 'Gem',
+  'cost.token_statistics': 'Tokenstatistik',
+  'logs.title': 'Live-logge',
+  'doctor.title': 'Systemdiagnostik',
+  'auth.pair_button': 'Par',
+  'auth.enter_code': 'Indtast engangskoden fra terminalen',
+  'auth.code_placeholder': '6-cifret kode',
+  'auth.pairing_progress': 'Parrer…',
+  'auth.logout': 'Log ud',
+  'common.languages': 'Sprog',
+  'common.select_language': 'Vælg sprog',
+  'header.dashboard_tagline': 'ZeroClaw-kontrolpanel',
+};
+
+export default da;
