+++ web/src/lib/i18n/locales/de.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const de: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Dashboard',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Werkzeuge',
+  'nav.cron': 'Geplante Aufgaben',
+  'nav.integrations': 'Integrationen',
+  'nav.memory': 'Speicher',
+  'nav.config': 'Konfiguration',
+  'nav.cost': 'Kosten',
+  'nav.logs': 'Protokolle',
+  'nav.doctor': 'Diagnose',
+  'dashboard.hero_title': 'Elektrisches Runtime-Dashboard',
+  'agent.placeholder': 'Nachricht eingeben…',
+  'tools.search': 'Werkzeuge suchen…',
+  'cron.add': 'Aufgabe hinzufügen',
+  'memory.add_memory': 'Speicher hinzufügen',
+  'config.save': 'Speichern',
+  'cost.token_statistics': 'Token-Statistiken',
+  'logs.title': 'Live-Protokolle',
+  'doctor.title': 'Systemdiagnose',
+  'auth.pair_button': 'Koppeln',
+  'auth.enter_code': 'Geben Sie den einmaligen Kopplungscode aus dem Terminal ein',
+  'auth.code_placeholder': '6-stelliger Code',
+  'auth.pairing_progress': 'Kopplung…',
+  'auth.logout': 'Abmelden',
+  'common.languages': 'Sprachen',
+  'common.select_language': 'Sprache auswählen',
+  'header.dashboard_tagline': 'ZeroClaw-Dashboard',
+};
+
+export default de;
