+++ web/src/lib/i18n/locales/el.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const el: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Πίνακας ελέγχου',
+  'nav.agent': 'Πράκτορας',
+  'nav.tools': 'Εργαλεία',
+  'nav.cron': 'Προγραμματισμένες εργασίες',
+  'nav.integrations': 'Ενσωματώσεις',
+  'nav.memory': 'Μνήμη',
+  'nav.config': 'Ρυθμίσεις',
+  'nav.cost': 'Κόστος',
+  'nav.logs': 'Αρχεία καταγραφής',
+  'nav.doctor': 'Διάγνωση',
+  'dashboard.hero_title': 'Ηλεκτρικός πίνακας runtime',
+  'agent.placeholder': 'Πληκτρολογήστε μήνυμα…',
+  'tools.search': 'Αναζήτηση εργαλείων…',
+  'cron.add': 'Προσθήκη εργασίας',
+  'memory.add_memory': 'Προσθήκη μνήμης',
+  'config.save': 'Αποθήκευση',
+  'cost.token_statistics': 'Στατιστικά token',
+  'logs.title': 'Ζωντανά αρχεία καταγραφής',
+  'doctor.title': 'Διάγνωση συστήματος',
+  'auth.pair_button': 'Σύζευξη',
+  'auth.enter_code': 'Εισαγάγετε τον εφάπαξ κωδικό σύζευξης από το terminal',
+  'auth.code_placeholder': '6ψήφιος κωδικός',
+  'auth.pairing_progress': 'Σύζευξη…',
+  'auth.logout': 'Αποσύνδεση',
+  'common.languages': 'Γλώσσες',
+  'common.select_language': 'Επιλέξτε γλώσσα',
+  'header.dashboard_tagline': 'Πίνακας ZeroClaw',
+};
+
+export default el;
