+++ web/src/lib/i18n/locales/en.ts
@@ -1,0 +1,273 @@
+const en = {
+  'nav.dashboard': 'Dashboard',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Tools',
+  'nav.cron': 'Scheduled Jobs',
+  'nav.integrations': 'Integrations',
+  'nav.memory': 'Memory',
+  'nav.config': 'Configuration',
+  'nav.cost': 'Cost Tracker',
+  'nav.logs': 'Logs',
+  'nav.doctor': 'Doctor',
+
+  'dashboard.title': 'Dashboard',
+  'dashboard.provider': 'Provider',
+  'dashboard.model': 'Model',
+  'dashboard.uptime': 'Uptime',
+  'dashboard.temperature': 'Temperature',
+  'dashboard.gateway_port': 'Gateway Port',
+  'dashboard.locale': 'Locale',
+  'dashboard.memory_backend': 'Memory Backend',
+  'dashboard.paired': 'Paired',
+  'dashboard.channels': 'Channels',
+  'dashboard.health': 'Health',
+  'dashboard.status': 'Status',
+  'dashboard.overview': 'Overview',
+  'dashboard.system_info': 'System Information',
+  'dashboard.quick_actions': 'Quick Actions',
+  'dashboard.load_failed': 'Dashboard load failed',
+  'dashboard.load_unknown_error': 'Unknown dashboard load error',
+  'dashboard.hero_eyebrow': 'ZeroClaw Command Deck',
+  'dashboard.hero_title': 'Electric Runtime Dashboard',
+  'dashboard.hero_subtitle': 'Real-time telemetry, cost pulse, and operations status in a single collapsible surface.',
+  'dashboard.live_gateway': 'Live Gateway',
+  'dashboard.unpaired': 'Unpaired',
+  'dashboard.provider_model': 'Provider / Model',
+  'dashboard.since_last_restart': 'Since last restart',
+  'dashboard.pairing_active': 'Pairing active',
+  'dashboard.no_paired_devices': 'No paired devices',
+  'dashboard.cost_pulse': 'Cost Pulse',
+  'dashboard.cost_subtitle': 'Session, daily, and monthly runtime spend',
+  'dashboard.session': 'Session',
+  'dashboard.daily': 'Daily',
+  'dashboard.monthly': 'Monthly',
+  'dashboard.channel_activity': 'Channel Activity',
+  'dashboard.channel_subtitle': 'Live integrations and route connectivity',
+  'dashboard.no_channels': 'No channels configured.',
+  'dashboard.active': 'Active',
+  'dashboard.inactive': 'Inactive',
+  'dashboard.component_health': 'Component Health',
+  'dashboard.component_subtitle': 'Runtime heartbeat and restart awareness',
+  'dashboard.no_component_health': 'No component health is currently available.',
+  'dashboard.restarts': 'Restarts',
+  'dashboard.unknown_provider': 'Unknown',
+
+  'agent.title': 'Agent Chat',
+  'agent.send': 'Send',
+  'agent.placeholder': 'Type a message...',
+  'agent.connecting': 'Connecting...',
+  'agent.connected': 'Connected',
+  'agent.disconnected': 'Disconnected',
+  'agent.reconnecting': 'Reconnecting...',
+  'agent.thinking': 'Thinking...',
+  'agent.tool_call': 'Tool Call',
+  'agent.tool_result': 'Tool Result',
+  'agent.connection_error': 'Connection error. Attempting to reconnect...',
+  'agent.failed_send': 'Failed to send message. Please try again.',
+  'agent.empty_title': 'ZeroClaw Agent',
+  'agent.empty_subtitle': 'Send a message to start the conversation',
+  'agent.unknown_error': 'Unknown error',
+
+  'tools.title': 'Available Tools',
+  'tools.name': 'Name',
+  'tools.description': 'Description',
+  'tools.parameters': 'Parameters',
+  'tools.search': 'Search tools...',
+  'tools.empty': 'No tools available.',
+  'tools.count': 'Total tools',
+  'tools.agent_tools': 'Agent Tools',
+  'tools.cli_tools': 'CLI Tools',
+  'tools.no_search_results': 'No tools match your search.',
+  'tools.parameter_schema': 'Parameter Schema',
+  'tools.path': 'Path',
+  'tools.version': 'Version',
+  'tools.load_failed': 'Failed to load tools',
+
+  'cron.title': 'Scheduled Jobs',
+  'cron.add': 'Add Job',
+  'cron.delete': 'Delete',
+  'cron.enable': 'Enable',
+  'cron.disable': 'Disable',
+  'cron.name': 'Name',
+  'cron.command': 'Command',
+  'cron.schedule': 'Schedule',
+  'cron.next_run': 'Next Run',
+  'cron.last_run': 'Last Run',
+  'cron.last_status': 'Last Status',
+  'cron.enabled': 'Enabled',
+  'cron.empty': 'No scheduled jobs.',
+  'cron.confirm_delete': 'Are you sure you want to delete this job?',
+  'cron.scheduled_tasks': 'Scheduled Tasks',
+  'cron.add_cron_job': 'Add Cron Job',
+  'cron.name_optional': 'Name (optional)',
+  'cron.schedule_required_command_required': 'Schedule and command are required.',
+  'cron.adding': 'Adding...',
+  'cron.no_tasks_configured': 'No scheduled tasks configured.',
+  'cron.load_failed': 'Failed to load cron jobs',
+  'cron.failed_add': 'Failed to add job',
+  'cron.failed_delete': 'Failed to delete job',
+  'cron.delete_prompt': 'Delete?',
+  'cron.id': 'ID',
+  'cron.disabled': 'Disabled',
+
+  'integrations.title': 'Integrations',
+  'integrations.available': 'Available',
+  'integrations.active': 'Active',
+  'integrations.coming_soon': 'Coming Soon',
+  'integrations.category': 'Category',
+  'integrations.status': 'Status',
+  'integrations.search': 'Search integrations...',
+  'integrations.empty': 'No integrations found.',
+  'integrations.activate': 'Activate',
+  'integrations.deactivate': 'Deactivate',
+  'integrations.load_failed': 'Failed to load integrations',
+  'integrations.all': 'all',
+
+  'memory.title': 'Memory Store',
+  'memory.search': 'Search memory...',
+  'memory.add': 'Store Memory',
+  'memory.delete': 'Delete',
+  'memory.key': 'Key',
+  'memory.content': 'Content',
+  'memory.category': 'Category',
+  'memory.timestamp': 'Timestamp',
+  'memory.session': 'Session',
+  'memory.score': 'Score',
+  'memory.empty': 'No memory entries found.',
+  'memory.confirm_delete': 'Are you sure you want to delete this memory entry?',
+  'memory.all_categories': 'All Categories',
+  'memory.add_memory': 'Add Memory',
+  'memory.search_entries': 'Search memory entries...',
+  'memory.load_failed': 'Failed to load memory',
+  'memory.key_content_required': 'Key and content are required.',
+  'memory.failed_store': 'Failed to store memory',
+  'memory.failed_delete': 'Failed to delete memory',
+  'memory.category_optional': 'Category (optional)',
+  'memory.key_placeholder': 'e.g. user_preferences',
+  'memory.content_placeholder': 'Memory content...',
+  'memory.category_placeholder': 'e.g. preferences, context, facts',
+  'memory.search_button': 'Search',
+  'memory.saving': 'Saving...',
+  'memory.delete_prompt': 'Delete?',
+
+  'config.title': 'Configuration',
+  'config.save': 'Save',
+  'config.reset': 'Reset',
+  'config.saved': 'Configuration saved successfully.',
+  'config.error': 'Failed to save configuration.',
+  'config.loading': 'Loading configuration...',
+  'config.editor_placeholder': 'TOML configuration...',
+  'config.saving': 'Saving...',
+  'config.masked_title': 'Sensitive fields are masked',
+  'config.masked_description': 'API keys, tokens, and passwords are hidden for security. To update a masked field, replace the entire masked value with your new value.',
+  'config.toml_configuration': 'TOML Configuration',
+  'config.lines': 'lines',
+
+  'cost.title': 'Cost Tracker',
+  'cost.session': 'Session Cost',
+  'cost.daily': 'Daily Cost',
+  'cost.monthly': 'Monthly Cost',
+  'cost.total_tokens': 'Total Tokens',
+  'cost.request_count': 'Requests',
+  'cost.by_model': 'Cost by Model',
+  'cost.model': 'Model',
+  'cost.tokens': 'Tokens',
+  'cost.requests': 'Requests',
+  'cost.usd': 'Cost (USD)',
+  'cost.total_requests': 'Total Requests',
+  'cost.token_statistics': 'Token Statistics',
+  'cost.avg_tokens_per_request': 'Avg Tokens / Request',
+  'cost.cost_per_1k_tokens': 'Cost per 1K Tokens',
+  'cost.model_breakdown': 'Model Breakdown',
+  'cost.no_model_data': 'No model data available.',
+  'cost.share': 'Share',
+  'cost.load_failed': 'Failed to load cost data',
+
+  'logs.title': 'Live Logs',
+  'logs.clear': 'Clear',
+  'logs.pause': 'Pause',
+  'logs.resume': 'Resume',
+  'logs.filter': 'Filter logs...',
+  'logs.empty': 'No log entries.',
+  'logs.connected': 'Connected to event stream.',
+  'logs.disconnected': 'Disconnected from event stream.',
+  'logs.events': 'events',
+  'logs.jump_to_bottom': 'Jump to bottom',
+  'logs.filter_label': 'Filter:',
+  'logs.paused_stream': 'Log streaming is paused.',
+  'logs.waiting_for_events': 'Waiting for events...',
+
+  'doctor.title': 'System Diagnostics',
+  'doctor.run': 'Run Diagnostics',
+  'doctor.running': 'Running diagnostics...',
+  'doctor.ok': 'OK',
+  'doctor.warn': 'Warning',
+  'doctor.error': 'Error',
+  'doctor.severity': 'Severity',
+  'doctor.category': 'Category',
+  'doctor.message': 'Message',
+  'doctor.empty': 'No diagnostics have been run yet.',
+  'doctor.summary': 'Diagnostic Summary',
+  'doctor.running_short': 'Running...',
+  'doctor.running_hint': 'This may take a few seconds.',
+  'doctor.issues_found': 'Issues Found',
+  'doctor.warnings': 'Warnings',
+  'doctor.all_clear': 'All Clear',
+  'doctor.instructions': 'Click "Run Diagnostics" to check your ZeroClaw installation.',
+
+  'auth.pair': 'Pair Device',
+  'auth.pairing_code': 'Pairing Code',
+  'auth.pair_button': 'Pair',
+  'auth.logout': 'Logout',
+  'auth.pairing_success': 'Pairing successful!',
+  'auth.pairing_failed': 'Pairing failed. Please try again.',
+  'auth.enter_code': 'Enter the one-time pairing code from your terminal',
+  'auth.code_placeholder': '6-digit code',
+  'auth.pairing_progress': 'Pairing...',
+
+  'common.loading': 'Loading...',
+  'common.error': 'An error occurred.',
+  'common.retry': 'Retry',
+  'common.cancel': 'Cancel',
+  'common.confirm': 'Confirm',
+  'common.save': 'Save',
+  'common.delete': 'Delete',
+  'common.edit': 'Edit',
+  'common.close': 'Close',
+  'common.yes': 'Yes',
+  'common.no': 'No',
+  'common.search': 'Search...',
+  'common.no_data': 'No data available.',
+  'common.refresh': 'Refresh',
+  'common.back': 'Back',
+  'common.actions': 'Actions',
+  'common.name': 'Name',
+  'common.description': 'Description',
+  'common.status': 'Status',
+  'common.created': 'Created',
+  'common.updated': 'Updated',
+  'common.languages': 'Languages',
+  'common.select_language': 'Select language',
+  'common.connecting': 'Connecting...',
+
+  'health.title': 'System Health',
+  'health.component': 'Component',
+  'health.status': 'Status',
+  'health.last_ok': 'Last OK',
+  'health.last_error': 'Last Error',
+  'health.restart_count': 'Restart Count',
+  'health.pid': 'Process ID',
+  'health.uptime': 'Uptime',
+  'health.updated_at': 'Updated At',
+
+  'header.dashboard_tagline': 'ZeroClaw dashboard',
+  'sidebar.gateway_dashboard': 'Gateway + Dashboard',
+  'sidebar.runtime_mode': 'Runtime Mode',
+  'navigation.open': 'Open navigation',
+  'navigation.close': 'Close navigation',
+  'navigation.expand': 'Expand navigation',
+  'navigation.collapse': 'Collapse navigation',
+} satisfies Record<string, string>;
+
+export type TranslationKeys = typeof en;
+export default en;
