+++ web/src/lib/i18n/locales/es.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const es: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Panel',
+  'nav.agent': 'Agente',
+  'nav.tools': 'Herramientas',
+  'nav.cron': 'Tareas programadas',
+  'nav.integrations': 'Integraciones',
+  'nav.memory': 'Memoria',
+  'nav.config': 'Configuración',
+  'nav.cost': 'Costos',
+  'nav.logs': 'Registros',
+  'nav.doctor': 'Diagnóstico',
+  'dashboard.hero_title': 'Panel eléctrico del runtime',
+  'agent.placeholder': 'Escribe un mensaje…',
+  'tools.search': 'Buscar herramientas…',
+  'cron.add': 'Agregar tarea',
+  'memory.add_memory': 'Agregar memoria',
+  'config.save': 'Guardar',
+  'cost.token_statistics': 'Estadísticas de tokens',
+  'logs.title': 'Registros en vivo',
+  'doctor.title': 'Diagnóstico del sistema',
+  'auth.pair_button': 'Vincular',
+  'auth.enter_code': 'Introduce el código de vinculación de un solo uso del terminal',
+  'auth.code_placeholder': 'Código de 6 dígitos',
+  'auth.pairing_progress': 'Vinculando…',
+  'auth.logout': 'Cerrar sesión',
+  'common.languages': 'Idiomas',
+  'common.select_language': 'Elegir idioma',
+  'header.dashboard_tagline': 'Panel de ZeroClaw',
+};
+
+export default es;
