+++ web/src/lib/i18n/locales/fi.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const fi: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Hallintapaneeli',
+  'nav.agent': 'Agentti',
+  'nav.tools': 'Työkalut',
+  'nav.cron': 'Ajastetut tehtävät',
+  'nav.integrations': 'Integraatiot',
+  'nav.memory': 'Muisti',
+  'nav.config': 'Asetukset',
+  'nav.cost': 'Kustannukset',
+  'nav.logs': 'Lokit',
+  'nav.doctor': 'Diagnostiikka',
+  'dashboard.hero_title': 'Sähköinen runtime-hallintapaneeli',
+  'agent.placeholder': 'Kirjoita viesti…',
+  'tools.search': 'Etsi työkaluja…',
+  'cron.add': 'Lisää tehtävä',
+  'memory.add_memory': 'Lisää muisti',
+  'config.save': 'Tallenna',
+  'cost.token_statistics': 'Token-tilastot',
+  'logs.title': 'Live-lokit',
+  'doctor.title': 'Järjestelmädiagnostiikka',
+  'auth.pair_button': 'Yhdistä',
+  'auth.enter_code': 'Syötä terminaalin kertakäyttöinen parituskoodi',
+  'auth.code_placeholder': '6-numeroinen koodi',
+  'auth.pairing_progress': 'Yhdistetään…',
+  'auth.logout': 'Kirjaudu ulos',
+  'common.languages': 'Kielet',
+  'common.select_language': 'Valitse kieli',
+  'header.dashboard_tagline': 'ZeroClaw-hallintapaneeli',
+};
+
+export default fi;
