+++ web/src/lib/i18n/locales/fr.ts
@@ -1,0 +1,51 @@
+import type { TranslationKeys } from './en';
+
+const fr: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Tableau de bord',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Outils',
+  'nav.cron': 'Tâches planifiées',
+  'nav.integrations': 'Intégrations',
+  'nav.memory': 'Mémoire',
+  'nav.config': 'Configuration',
+  'nav.cost': 'Coûts',
+  'nav.logs': 'Journaux',
+  'nav.doctor': 'Diagnostic',
+  'dashboard.hero_title': 'Tableau de bord runtime électrique',
+  'dashboard.live_gateway': 'Passerelle active',
+  'dashboard.unpaired': 'Non appairé',
+  'agent.title': 'Chat agent',
+  'agent.placeholder': 'Saisissez un message…',
+  'agent.connecting': 'Connexion…',
+  'agent.connected': 'Connecté',
+  'agent.disconnected': 'Déconnecté',
+  'tools.search': 'Rechercher des outils…',
+  'tools.agent_tools': 'Outils agent',
+  'tools.cli_tools': 'Outils CLI',
+  'cron.add': 'Ajouter une tâche',
+  'cron.scheduled_tasks': 'Tâches planifiées',
+  'integrations.title': 'Intégrations',
+  'memory.add_memory': 'Ajouter une mémoire',
+  'memory.search_entries': 'Rechercher dans la mémoire…',
+  'config.save': 'Enregistrer',
+  'config.saving': 'Enregistrement…',
+  'cost.session': 'Coût de session',
+  'cost.daily': 'Coût journalier',
+  'cost.monthly': 'Coût mensuel',
+  'logs.title': 'Journaux en direct',
+  'logs.pause': 'Pause',
+  'logs.resume': 'Reprendre',
+  'doctor.title': 'Diagnostic système',
+  'doctor.run': 'Lancer le diagnostic',
+  'doctor.running_short': 'Exécution…',
+  'auth.pair_button': 'Associer',
+  'auth.enter_code': 'Entrez le code d\u2019appairage à usage unique affiché dans le terminal',
+  'auth.code_placeholder': 'Code à 6 chiffres',
+  'auth.pairing_progress': 'Appairage…',
+  'auth.logout': 'Déconnexion',
+  'common.languages': 'Langues',
+  'common.select_language': 'Choisir la langue',
+  'header.dashboard_tagline': 'Tableau de bord ZeroClaw',
+};
+
+export default fr;
