+++ web/src/lib/i18n/locales/he.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const he: Partial<TranslationKeys> = {
+  'nav.dashboard': 'לוח מחוונים',
+  'nav.agent': 'סוכן',
+  'nav.tools': 'כלים',
+  'nav.cron': 'משימות מתוזמנות',
+  'nav.integrations': 'אינטגרציות',
+  'nav.memory': 'זיכרון',
+  'nav.config': 'תצורה',
+  'nav.cost': 'מעקב עלויות',
+  'nav.logs': 'יומנים',
+  'nav.doctor': 'אבחון',
+  'dashboard.hero_title': 'לוח מחוונים חשמלי של זמן הריצה',
+  'agent.placeholder': 'הקלד הודעה…',
+  'tools.search': 'חפש כלים…',
+  'cron.add': 'הוסף משימה',
+  'memory.add_memory': 'הוסף זיכרון',
+  'config.save': 'שמור',
+  'cost.token_statistics': 'סטטיסטיקות אסימונים',
+  'logs.title': 'יומנים חיים',
+  'doctor.title': 'אבחון מערכת',
+  'auth.pair_button': 'התאמה',
+  'auth.enter_code': 'הזן את קוד ההתאמה החד-פעמי מהמסוף',
+  'auth.code_placeholder': 'קוד בן 6 ספרות',
+  'auth.pairing_progress': 'מתבצעת התאמה…',
+  'auth.logout': 'התנתק',
+  'common.languages': 'שפות',
+  'common.select_language': 'בחר שפה',
+  'header.dashboard_tagline': 'לוח המחוונים של ZeroClaw',
+};
+
+export default he;
