+++ web/src/lib/i18n/locales/hi.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const hi: Partial<TranslationKeys> = {
+  'nav.dashboard': 'डैशबोर्ड',
+  'nav.agent': 'एजेंट',
+  'nav.tools': 'टूल्स',
+  'nav.cron': 'निर्धारित कार्य',
+  'nav.integrations': 'इंटीग्रेशन',
+  'nav.memory': 'मेमोरी',
+  'nav.config': 'कॉन्फ़िगरेशन',
+  'nav.cost': 'लागत ट्रैकर',
+  'nav.logs': 'लॉग्स',
+  'nav.doctor': 'जाँच',
+  'dashboard.hero_title': 'इलेक्ट्रिक रनटाइम डैशबोर्ड',
+  'agent.placeholder': 'संदेश लिखें…',
+  'tools.search': 'टूल खोजें…',
+  'cron.add': 'कार्य जोड़ें',
+  'memory.add_memory': 'मेमोरी जोड़ें',
+  'config.save': 'सहेजें',
+  'cost.token_statistics': 'टोकन आँकड़े',
+  'logs.title': 'लाइव लॉग्स',
+  'doctor.title': 'सिस्टम जाँच',
+  'auth.pair_button': 'पेयर करें',
+  'auth.enter_code': 'टर्मिनल से एक-बार वाला पेयरिंग कोड दर्ज करें',
+  'auth.code_placeholder': '6-अंकों का कोड',
+  'auth.pairing_progress': 'पेयर किया जा रहा है…',
+  'auth.logout': 'लॉग आउट',
+  'common.languages': 'भाषाएँ',
+  'common.select_language': 'भाषा चुनें',
+  'header.dashboard_tagline': 'ZeroClaw डैशबोर्ड',
+};
+
+export default hi;
