+++ web/src/lib/i18n/locales/hu.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const hu: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Irányítópult',
+  'nav.agent': 'Ügynök',
+  'nav.tools': 'Eszközök',
+  'nav.cron': 'Ütemezett feladatok',
+  'nav.integrations': 'Integrációk',
+  'nav.memory': 'Memória',
+  'nav.config': 'Konfiguráció',
+  'nav.cost': 'Költségek',
+  'nav.logs': 'Naplók',
+  'nav.doctor': 'Diagnosztika',
+  'dashboard.hero_title': 'Elektromos runtime irányítópult',
+  'agent.placeholder': 'Írjon üzenetet…',
+  'tools.search': 'Eszközök keresése…',
+  'cron.add': 'Feladat hozzáadása',
+  'memory.add_memory': 'Memória hozzáadása',
+  'config.save': 'Mentés',
+  'cost.token_statistics': 'Tokenstatisztika',
+  'logs.title': 'Élő naplók',
+  'doctor.title': 'Rendszerdiagnosztika',
+  'auth.pair_button': 'Párosítás',
+  'auth.enter_code': 'Adja meg a terminál egyszer használatos párosítási kódját',
+  'auth.code_placeholder': '6 számjegyű kód',
+  'auth.pairing_progress': 'Párosítás…',
+  'auth.logout': 'Kijelentkezés',
+  'common.languages': 'Nyelvek',
+  'common.select_language': 'Nyelv kiválasztása',
+  'header.dashboard_tagline': 'ZeroClaw irányítópult',
+};
+
+export default hu;
