+++ web/src/lib/i18n/locales/id.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const id: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Dasbor',
+  'nav.agent': 'Agen',
+  'nav.tools': 'Alat',
+  'nav.cron': 'Tugas terjadwal',
+  'nav.integrations': 'Integrasi',
+  'nav.memory': 'Memori',
+  'nav.config': 'Konfigurasi',
+  'nav.cost': 'Biaya',
+  'nav.logs': 'Log',
+  'nav.doctor': 'Diagnosis',
+  'dashboard.hero_title': 'Dasbor runtime elektrik',
+  'agent.placeholder': 'Tulis pesan…',
+  'tools.search': 'Cari alat…',
+  'cron.add': 'Tambah tugas',
+  'memory.add_memory': 'Tambah memori',
+  'config.save': 'Simpan',
+  'cost.token_statistics': 'Statistik token',
+  'logs.title': 'Log langsung',
+  'doctor.title': 'Diagnosis sistem',
+  'auth.pair_button': 'Pasangkan',
+  'auth.enter_code': 'Masukkan kode pairing sekali pakai dari terminal',
+  'auth.code_placeholder': 'Kode 6 digit',
+  'auth.pairing_progress': 'Sedang memasangkan…',
+  'auth.logout': 'Keluar',
+  'common.languages': 'Bahasa',
+  'common.select_language': 'Pilih bahasa',
+  'header.dashboard_tagline': 'Dasbor ZeroClaw',
+};
+
+export default id;
