+++ web/src/lib/i18n/locales/it.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const it: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Dashboard',
+  'nav.agent': 'Agente',
+  'nav.tools': 'Strumenti',
+  'nav.cron': 'Attività pianificate',
+  'nav.integrations': 'Integrazioni',
+  'nav.memory': 'Memoria',
+  'nav.config': 'Configurazione',
+  'nav.cost': 'Costi',
+  'nav.logs': 'Log',
+  'nav.doctor': 'Diagnostica',
+  'dashboard.hero_title': 'Dashboard runtime elettrica',
+  'agent.placeholder': 'Scrivi un messaggio…',
+  'tools.search': 'Cerca strumenti…',
+  'cron.add': 'Aggiungi attività',
+  'memory.add_memory': 'Aggiungi memoria',
+  'config.save': 'Salva',
+  'cost.token_statistics': 'Statistiche token',
+  'logs.title': 'Log in tempo reale',
+  'doctor.title': 'Diagnostica di sistema',
+  'auth.pair_button': 'Associa',
+  'auth.enter_code': 'Inserisci il codice di associazione monouso dal terminale',
+  'auth.code_placeholder': 'Codice a 6 cifre',
+  'auth.pairing_progress': 'Associazione…',
+  'auth.logout': 'Disconnetti',
+  'common.languages': 'Lingue',
+  'common.select_language': 'Scegli lingua',
+  'header.dashboard_tagline': 'Dashboard di ZeroClaw',
+};
+
+export default it;
