+++ web/src/lib/i18n/locales/ja.ts
@@ -1,0 +1,51 @@
+import type { TranslationKeys } from './en';
+
+const ja: Partial<TranslationKeys> = {
+  'nav.dashboard': 'ダッシュボード',
+  'nav.agent': 'エージェント',
+  'nav.tools': 'ツール',
+  'nav.cron': 'スケジュール',
+  'nav.integrations': '連携',
+  'nav.memory': 'メモリ',
+  'nav.config': '設定',
+  'nav.cost': 'コスト',
+  'nav.logs': 'ログ',
+  'nav.doctor': '診断',
+  'dashboard.hero_title': 'エレクトリック・ランタイム・ダッシュボード',
+  'dashboard.live_gateway': 'ライブゲートウェイ',
+  'dashboard.unpaired': '未ペア',
+  'agent.title': 'エージェントチャット',
+  'agent.placeholder': 'メッセージを入力…',
+  'agent.connecting': '接続中…',
+  'agent.connected': '接続済み',
+  'agent.disconnected': '切断済み',
+  'tools.search': 'ツールを検索…',
+  'tools.agent_tools': 'エージェントツール',
+  'tools.cli_tools': 'CLI ツール',
+  'cron.add': 'ジョブを追加',
+  'cron.scheduled_tasks': 'スケジュールされたジョブ',
+  'integrations.title': '連携',
+  'memory.add_memory': 'メモリを追加',
+  'memory.search_entries': 'メモリエントリを検索…',
+  'config.save': '保存',
+  'config.saving': '保存中…',
+  'cost.session': 'セッションコスト',
+  'cost.daily': '日次コスト',
+  'cost.monthly': '月次コスト',
+  'logs.title': 'ライブログ',
+  'logs.pause': '一時停止',
+  'logs.resume': '再開',
+  'doctor.title': 'システム診断',
+  'doctor.run': '診断を実行',
+  'doctor.running_short': '実行中…',
+  'auth.pair_button': 'ペアリング',
+  'auth.enter_code': '端末のワンタイムペアリングコードを入力してください',
+  'auth.code_placeholder': '6桁のコード',
+  'auth.pairing_progress': 'ペアリング中…',
+  'auth.logout': 'ログアウト',
+  'common.languages': '言語',
+  'common.select_language': '言語を選択',
+  'header.dashboard_tagline': 'ZeroClaw ダッシュボード',
+};
+
+export default ja;
