+++ web/src/lib/i18n/locales/ko.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const ko: Partial<TranslationKeys> = {
+  'nav.dashboard': '대시보드',
+  'nav.agent': '에이전트',
+  'nav.tools': '도구',
+  'nav.cron': '예약 작업',
+  'nav.integrations': '통합',
+  'nav.memory': '메모리',
+  'nav.config': '설정',
+  'nav.cost': '비용 추적',
+  'nav.logs': '로그',
+  'nav.doctor': '진단',
+  'dashboard.hero_title': '전기 런타임 대시보드',
+  'agent.placeholder': '메시지를 입력하세요…',
+  'tools.search': '도구 검색…',
+  'cron.add': '작업 추가',
+  'memory.add_memory': '메모리 추가',
+  'config.save': '저장',
+  'cost.token_statistics': '토큰 통계',
+  'logs.title': '실시간 로그',
+  'doctor.title': '시스템 진단',
+  'auth.pair_button': '페어링',
+  'auth.enter_code': '터미널에 표시된 일회용 페어링 코드를 입력하세요',
+  'auth.code_placeholder': '6자리 코드',
+  'auth.pairing_progress': '페어링 중…',
+  'auth.logout': '로그아웃',
+  'common.languages': '언어',
+  'common.select_language': '언어 선택',
+  'header.dashboard_tagline': 'ZeroClaw 대시보드',
+};
+
+export default ko;
