+++ web/src/lib/i18n/locales/nb.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const nb: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Kontrollpanel',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Verktøy',
+  'nav.cron': 'Planlagte jobber',
+  'nav.integrations': 'Integrasjoner',
+  'nav.memory': 'Minne',
+  'nav.config': 'Konfigurasjon',
+  'nav.cost': 'Kostnader',
+  'nav.logs': 'Logger',
+  'nav.doctor': 'Diagnostikk',
+  'dashboard.hero_title': 'Elektrisk runtime-kontrollpanel',
+  'agent.placeholder': 'Skriv en melding…',
+  'tools.search': 'Søk etter verktøy…',
+  'cron.add': 'Legg til jobb',
+  'memory.add_memory': 'Legg til minne',
+  'config.save': 'Lagre',
+  'cost.token_statistics': 'Tokenstatistikk',
+  'logs.title': 'Live-logger',
+  'doctor.title': 'Systemdiagnostikk',
+  'auth.pair_button': 'Koble til',
+  'auth.enter_code': 'Skriv inn engangskoden fra terminalen',
+  'auth.code_placeholder': '6-sifret kode',
+  'auth.pairing_progress': 'Kobler til…',
+  'auth.logout': 'Logg ut',
+  'common.languages': 'Språk',
+  'common.select_language': 'Velg språk',
+  'header.dashboard_tagline': 'ZeroClaw-kontrollpanel',
+};
+
+export default nb;
