+++ web/src/lib/i18n/locales/nl.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const nl: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Dashboard',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Tools',
+  'nav.cron': 'Geplande taken',
+  'nav.integrations': 'Integraties',
+  'nav.memory': 'Geheugen',
+  'nav.config': 'Configuratie',
+  'nav.cost': 'Kosten',
+  'nav.logs': 'Logs',
+  'nav.doctor': 'Diagnose',
+  'dashboard.hero_title': 'Elektrisch runtime-dashboard',
+  'agent.placeholder': 'Typ een bericht…',
+  'tools.search': 'Tools zoeken…',
+  'cron.add': 'Taak toevoegen',
+  'memory.add_memory': 'Geheugen toevoegen',
+  'config.save': 'Opslaan',
+  'cost.token_statistics': 'Tokenstatistieken',
+  'logs.title': 'Live-logs',
+  'doctor.title': 'Systeemdiagnose',
+  'auth.pair_button': 'Koppelen',
+  'auth.enter_code': 'Voer de eenmalige koppelcode uit de terminal in',
+  'auth.code_placeholder': '6-cijferige code',
+  'auth.pairing_progress': 'Koppelen…',
+  'auth.logout': 'Afmelden',
+  'common.languages': 'Talen',
+  'common.select_language': 'Kies taal',
+  'header.dashboard_tagline': 'ZeroClaw-dashboard',
+};
+
+export default nl;
