+++ web/src/lib/i18n/locales/pl.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const pl: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Pulpit',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Narzędzia',
+  'nav.cron': 'Zaplanowane zadania',
+  'nav.integrations': 'Integracje',
+  'nav.memory': 'Pamięć',
+  'nav.config': 'Konfiguracja',
+  'nav.cost': 'Koszty',
+  'nav.logs': 'Logi',
+  'nav.doctor': 'Diagnostyka',
+  'dashboard.hero_title': 'Elektryczny pulpit runtime',
+  'agent.placeholder': 'Wpisz wiadomość…',
+  'tools.search': 'Szukaj narzędzi…',
+  'cron.add': 'Dodaj zadanie',
+  'memory.add_memory': 'Dodaj pamięć',
+  'config.save': 'Zapisz',
+  'cost.token_statistics': 'Statystyki tokenów',
+  'logs.title': 'Logi na żywo',
+  'doctor.title': 'Diagnostyka systemu',
+  'auth.pair_button': 'Sparuj',
+  'auth.enter_code': 'Wprowadź jednorazowy kod parowania z terminala',
+  'auth.code_placeholder': '6-cyfrowy kod',
+  'auth.pairing_progress': 'Parowanie…',
+  'auth.logout': 'Wyloguj',
+  'common.languages': 'Języki',
+  'common.select_language': 'Wybierz język',
+  'header.dashboard_tagline': 'Pulpit ZeroClaw',
+};
+
+export default pl;
