+++ web/src/lib/i18n/locales/pt.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const pt: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Painel',
+  'nav.agent': 'Agente',
+  'nav.tools': 'Ferramentas',
+  'nav.cron': 'Tarefas agendadas',
+  'nav.integrations': 'Integrações',
+  'nav.memory': 'Memória',
+  'nav.config': 'Configuração',
+  'nav.cost': 'Custos',
+  'nav.logs': 'Logs',
+  'nav.doctor': 'Diagnóstico',
+  'dashboard.hero_title': 'Painel elétrico do runtime',
+  'agent.placeholder': 'Digite uma mensagem…',
+  'tools.search': 'Buscar ferramentas…',
+  'cron.add': 'Adicionar tarefa',
+  'memory.add_memory': 'Adicionar memória',
+  'config.save': 'Salvar',
+  'cost.token_statistics': 'Estatísticas de tokens',
+  'logs.title': 'Logs ao vivo',
+  'doctor.title': 'Diagnóstico do sistema',
+  'auth.pair_button': 'Parear',
+  'auth.enter_code': 'Digite o código único de pareamento mostrado no terminal',
+  'auth.code_placeholder': 'Código de 6 dígitos',
+  'auth.pairing_progress': 'Pareando…',
+  'auth.logout': 'Sair',
+  'common.languages': 'Idiomas',
+  'common.select_language': 'Escolher idioma',
+  'header.dashboard_tagline': 'Painel do ZeroClaw',
+};
+
+export default pt;
