+++ web/src/lib/i18n/locales/ro.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const ro: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Tablou de bord',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Unelte',
+  'nav.cron': 'Sarcini programate',
+  'nav.integrations': 'Integrări',
+  'nav.memory': 'Memorie',
+  'nav.config': 'Configurație',
+  'nav.cost': 'Costuri',
+  'nav.logs': 'Jurnale',
+  'nav.doctor': 'Diagnostic',
+  'dashboard.hero_title': 'Tablou de bord runtime electric',
+  'agent.placeholder': 'Scrie un mesaj…',
+  'tools.search': 'Caută unelte…',
+  'cron.add': 'Adaugă sarcină',
+  'memory.add_memory': 'Adaugă memorie',
+  'config.save': 'Salvează',
+  'cost.token_statistics': 'Statistici tokenuri',
+  'logs.title': 'Jurnale live',
+  'doctor.title': 'Diagnostic sistem',
+  'auth.pair_button': 'Asociază',
+  'auth.enter_code': 'Introdu codul unic de asociere din terminal',
+  'auth.code_placeholder': 'Cod din 6 cifre',
+  'auth.pairing_progress': 'Asociere…',
+  'auth.logout': 'Deconectare',
+  'common.languages': 'Limbi',
+  'common.select_language': 'Alege limba',
+  'header.dashboard_tagline': 'Tabloul ZeroClaw',
+};
+
+export default ro;
