+++ web/src/lib/i18n/locales/ru.ts
@@ -1,0 +1,51 @@
+import type { TranslationKeys } from './en';
+
+const ru: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Панель',
+  'nav.agent': 'Агент',
+  'nav.tools': 'Инструменты',
+  'nav.cron': 'Задания',
+  'nav.integrations': 'Интеграции',
+  'nav.memory': 'Память',
+  'nav.config': 'Конфигурация',
+  'nav.cost': 'Расходы',
+  'nav.logs': 'Логи',
+  'nav.doctor': 'Диагностика',
+  'dashboard.hero_title': 'Панель электрического рантайма',
+  'dashboard.live_gateway': 'Живой шлюз',
+  'dashboard.unpaired': 'Не сопряжено',
+  'agent.title': 'Чат агента',
+  'agent.placeholder': 'Введите сообщение…',
+  'agent.connecting': 'Подключение…',
+  'agent.connected': 'Подключено',
+  'agent.disconnected': 'Отключено',
+  'tools.search': 'Поиск инструментов…',
+  'tools.agent_tools': 'Инструменты агента',
+  'tools.cli_tools': 'CLI-инструменты',
+  'cron.add': 'Добавить задачу',
+  'cron.scheduled_tasks': 'Запланированные задания',
+  'integrations.title': 'Интеграции',
+  'memory.add_memory': 'Добавить память',
+  'memory.search_entries': 'Искать записи памяти…',
+  'config.save': 'Сохранить',
+  'config.saving': 'Сохранение…',
+  'cost.session': 'Стоимость сессии',
+  'cost.daily': 'Стоимость за день',
+  'cost.monthly': 'Стоимость за месяц',
+  'logs.title': 'Живые логи',
+  'logs.pause': 'Пауза',
+  'logs.resume': 'Продолжить',
+  'doctor.title': 'Диагностика системы',
+  'doctor.run': 'Запустить диагностику',
+  'doctor.running_short': 'Выполняется…',
+  'auth.pair_button': 'Сопрячь',
+  'auth.enter_code': 'Введите одноразовый код сопряжения из терминала',
+  'auth.code_placeholder': '6-значный код',
+  'auth.pairing_progress': 'Сопряжение…',
+  'auth.logout': 'Выйти',
+  'common.languages': 'Языки',
+  'common.select_language': 'Выберите язык',
+  'header.dashboard_tagline': 'Панель ZeroClaw',
+};
+
+export default ru;
