+++ web/src/lib/i18n/locales/sv.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const sv: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Instrumentpanel',
+  'nav.agent': 'Agent',
+  'nav.tools': 'Verktyg',
+  'nav.cron': 'Schemalagda jobb',
+  'nav.integrations': 'Integrationer',
+  'nav.memory': 'Minne',
+  'nav.config': 'Konfiguration',
+  'nav.cost': 'Kostnader',
+  'nav.logs': 'Loggar',
+  'nav.doctor': 'Diagnostik',
+  'dashboard.hero_title': 'Elektrisk runtimepanel',
+  'agent.placeholder': 'Skriv ett meddelande…',
+  'tools.search': 'Sök verktyg…',
+  'cron.add': 'Lägg till jobb',
+  'memory.add_memory': 'Lägg till minne',
+  'config.save': 'Spara',
+  'cost.token_statistics': 'Tokenstatistik',
+  'logs.title': 'Live-loggar',
+  'doctor.title': 'Systemdiagnostik',
+  'auth.pair_button': 'Para',
+  'auth.enter_code': 'Ange engångskoden från terminalen',
+  'auth.code_placeholder': '6-siffrig kod',
+  'auth.pairing_progress': 'Parar…',
+  'auth.logout': 'Logga ut',
+  'common.languages': 'Språk',
+  'common.select_language': 'Välj språk',
+  'header.dashboard_tagline': 'ZeroClaw-panel',
+};
+
+export default sv;
