+++ web/src/lib/i18n/locales/th.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const th: Partial<TranslationKeys> = {
+  'nav.dashboard': 'แดชบอร์ด',
+  'nav.agent': 'เอเจนต์',
+  'nav.tools': 'เครื่องมือ',
+  'nav.cron': 'งานที่ตั้งเวลา',
+  'nav.integrations': 'การเชื่อมต่อ',
+  'nav.memory': 'หน่วยความจำ',
+  'nav.config': 'การกำหนดค่า',
+  'nav.cost': 'ต้นทุน',
+  'nav.logs': 'บันทึก',
+  'nav.doctor': 'วินิจฉัย',
+  'dashboard.hero_title': 'แดชบอร์ดรันไทม์ไฟฟ้า',
+  'agent.placeholder': 'พิมพ์ข้อความ…',
+  'tools.search': 'ค้นหาเครื่องมือ…',
+  'cron.add': 'เพิ่มงาน',
+  'memory.add_memory': 'เพิ่มหน่วยความจำ',
+  'config.save': 'บันทึก',
+  'cost.token_statistics': 'สถิติโทเค็น',
+  'logs.title': 'บันทึกสด',
+  'doctor.title': 'วินิจฉัยระบบ',
+  'auth.pair_button': 'จับคู่',
+  'auth.enter_code': 'ป้อนรหัสจับคู่แบบใช้ครั้งเดียวจากเทอร์มินัล',
+  'auth.code_placeholder': 'รหัส 6 หลัก',
+  'auth.pairing_progress': 'กำลังจับคู่…',
+  'auth.logout': 'ออกจากระบบ',
+  'common.languages': 'ภาษา',
+  'common.select_language': 'เลือกภาษา',
+  'header.dashboard_tagline': 'แดชบอร์ด ZeroClaw',
+};
+
+export default th;
