+++ web/src/lib/i18n/locales/tl.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const tl: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Dashboard',
+  'nav.agent': 'Ahente',
+  'nav.tools': 'Mga Tool',
+  'nav.cron': 'Naka-iskedyul na Trabaho',
+  'nav.integrations': 'Mga Integrasyon',
+  'nav.memory': 'Alaala',
+  'nav.config': 'Konpigurasyon',
+  'nav.cost': 'Pagsubaybay sa Gastos',
+  'nav.logs': 'Mga Log',
+  'nav.doctor': 'Diyagnostiko',
+  'dashboard.hero_title': 'Elektrikong Dashboard ng Runtime',
+  'agent.placeholder': 'Mag-type ng mensahe…',
+  'tools.search': 'Maghanap ng tool…',
+  'cron.add': 'Magdagdag ng gawain',
+  'memory.add_memory': 'Magdagdag ng alaala',
+  'config.save': 'I-save',
+  'cost.token_statistics': 'Estadistika ng Token',
+  'logs.title': 'Mga Live Log',
+  'doctor.title': 'Diyagnostiko ng System',
+  'auth.pair_button': 'Ipares',
+  'auth.enter_code': 'Ilagay ang isang beses na pairing code mula sa terminal',
+  'auth.code_placeholder': '6-digit na code',
+  'auth.pairing_progress': 'Pinapares…',
+  'auth.logout': 'Mag-logout',
+  'common.languages': 'Mga Wika',
+  'common.select_language': 'Piliin ang wika',
+  'header.dashboard_tagline': 'Dashboard ng ZeroClaw',
+};
+
+export default tl;
