+++ web/src/lib/i18n/locales/tr.ts
@@ -1,0 +1,157 @@
+import type { TranslationKeys } from './en';
+
+const tr: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Kontrol Paneli',
+  'nav.agent': 'Ajan',
+  'nav.tools': 'Araçlar',
+  'nav.cron': 'Zamanlanmış Görevler',
+  'nav.integrations': 'Entegrasyonlar',
+  'nav.memory': 'Hafıza',
+  'nav.config': 'Yapılandırma',
+  'nav.cost': 'Maliyet Takibi',
+  'nav.logs': 'Kayıtlar',
+  'nav.doctor': 'Doktor',
+  'agent.title': 'Ajan Sohbeti',
+  'agent.send': 'Gönder',
+  'agent.placeholder': 'Bir mesaj yazın...',
+  'agent.connecting': 'Bağlanıyor...',
+  'agent.connected': 'Bağlı',
+  'agent.disconnected': 'Bağlantı Kesildi',
+  'agent.reconnecting': 'Yeniden bağlanıyor...',
+  'agent.thinking': 'Düşünüyor...',
+  'agent.tool_call': 'Araç Çağrısı',
+  'agent.tool_result': 'Araç Sonucu',
+  'agent.connection_error': 'Bağlantı hatası. Yeniden bağlanmaya çalışılıyor...',
+  'agent.failed_send': 'Mesaj gönderilemedi. Lütfen tekrar deneyin.',
+  'agent.empty_title': 'ZeroClaw Ajanı',
+  'agent.empty_subtitle': 'Konuşmayı başlatmak için bir mesaj gönderin',
+  'dashboard.title': 'Kontrol Paneli',
+  'dashboard.provider': 'Sağlayıcı',
+  'dashboard.model': 'Model',
+  'dashboard.uptime': 'Çalışma Süresi',
+  'dashboard.temperature': 'Sıcaklık',
+  'dashboard.gateway_port': 'Ağ Geçidi Portu',
+  'dashboard.locale': 'Yerel Ayar',
+  'dashboard.memory_backend': 'Hafıza Arka Ucu',
+  'dashboard.hero_eyebrow': 'ZeroClaw Komuta Güvertesi',
+  'dashboard.hero_title': 'Elektrik Çalışma Zamanı Paneli',
+  'dashboard.hero_subtitle': 'Gerçek zamanlı telemetri, maliyet akışı ve operasyon durumunu tek bir daraltılabilir yüzeyde görün.',
+  'dashboard.live_gateway': 'Canlı Ağ Geçidi',
+  'dashboard.unpaired': 'Eşleşmemiş',
+  'dashboard.provider_model': 'Sağlayıcı / Model',
+  'dashboard.since_last_restart': 'Son yeniden başlatmadan beri',
+  'dashboard.pairing_active': 'Eşleştirme etkin',
+  'dashboard.no_paired_devices': 'Eşleşmiş cihaz yok',
+  'dashboard.cost_pulse': 'Maliyet Nabzı',
+  'dashboard.cost_subtitle': 'Oturum, günlük ve aylık çalışma zamanı harcaması',
+  'dashboard.session': 'Oturum',
+  'dashboard.daily': 'Günlük',
+  'dashboard.monthly': 'Aylık',
+  'dashboard.channel_activity': 'Kanal Etkinliği',
+  'dashboard.channel_subtitle': 'Canlı entegrasyonlar ve rota bağlantısı',
+  'dashboard.no_channels': 'Hiç kanal yapılandırılmamış.',
+  'dashboard.active': 'Aktif',
+  'dashboard.inactive': 'Pasif',
+  'dashboard.component_health': 'Bileşen Sağlığı',
+  'dashboard.component_subtitle': 'Çalışma zamanı nabzı ve yeniden başlatma farkındalığı',
+  'dashboard.no_component_health': 'Şu anda bileşen sağlığı bilgisi yok.',
+  'dashboard.restarts': 'Yeniden Başlatmalar',
+  'tools.title': 'Mevcut Araçlar',
+  'tools.search': 'Araç ara...',
+  'tools.agent_tools': 'Ajan Araçları',
+  'tools.cli_tools': 'CLI Araçları',
+  'tools.no_search_results': 'Aramanızla eşleşen araç yok.',
+  'tools.parameter_schema': 'Parametre Şeması',
+  'tools.path': 'Yol',
+  'tools.version': 'Sürüm',
+  'cron.title': 'Zamanlanmış Görevler',
+  'cron.add': 'Görev Ekle',
+  'cron.scheduled_tasks': 'Zamanlanmış Görevler',
+  'cron.add_cron_job': 'Cron Görevi Ekle',
+  'cron.name_optional': 'Ad (isteğe bağlı)',
+  'cron.schedule_required_command_required': 'Zamanlama ve komut gereklidir.',
+  'cron.adding': 'Ekleniyor...',
+  'cron.no_tasks_configured': 'Zamanlanmış görev yapılandırılmamış.',
+  'cron.load_failed': 'Cron görevleri yüklenemedi',
+  'cron.failed_add': 'Görev eklenemedi',
+  'cron.failed_delete': 'Görev silinemedi',
+  'cron.delete_prompt': 'Silinsin mi?',
+  'cron.disabled': 'Devre Dışı',
+  'integrations.title': 'Entegrasyonlar',
+  'integrations.available': 'Mevcut',
+  'integrations.active': 'Aktif',
+  'integrations.coming_soon': 'Yakında',
+  'integrations.empty': 'Entegrasyon bulunamadı.',
+  'integrations.load_failed': 'Entegrasyonlar yüklenemedi',
+  'memory.title': 'Hafıza Deposu',
+  'memory.add_memory': 'Hafıza Ekle',
+  'memory.search_entries': 'Hafıza girdilerinde ara...',
+  'memory.all_categories': 'Tüm Kategoriler',
+  'memory.search_button': 'Ara',
+  'memory.load_failed': 'Hafıza yüklenemedi',
+  'memory.key_content_required': 'Anahtar ve içerik gereklidir.',
+  'memory.failed_store': 'Hafıza kaydedilemedi',
+  'memory.failed_delete': 'Hafıza silinemedi',
+  'memory.category_optional': 'Kategori (isteğe bağlı)',
+  'memory.key_placeholder': 'örn. kullanici_tercihleri',
+  'memory.content_placeholder': 'Hafıza içeriği...',
+  'memory.category_placeholder': 'örn. tercihler, bağlam, gerçekler',
+  'memory.saving': 'Kaydediliyor...',
+  'memory.delete_prompt': 'Silinsin mi?',
+  'config.title': 'Yapılandırma',
+  'config.save': 'Kaydet',
+  'config.saved': 'Yapılandırma başarıyla kaydedildi.',
+  'config.saving': 'Kaydediliyor...',
+  'config.masked_title': 'Hassas alanlar maskelendi',
+  'config.masked_description': 'Güvenlik için API anahtarları, belirteçler ve parolalar gizlenir. Maskelenmiş bir alanı güncellemek için tüm maskeli değeri yeni değerinizle değiştirin.',
+  'config.toml_configuration': 'TOML Yapılandırması',
+  'config.lines': 'satır',
+  'cost.title': 'Maliyet Takibi',
+  'cost.session': 'Oturum Maliyeti',
+  'cost.daily': 'Günlük Maliyet',
+  'cost.monthly': 'Aylık Maliyet',
+  'cost.total_requests': 'Toplam İstek',
+  'cost.token_statistics': 'Belirteç İstatistikleri',
+  'cost.avg_tokens_per_request': 'İstek Başına Ort. Belirteç',
+  'cost.cost_per_1k_tokens': '1K Belirteç Başına Maliyet',
+  'cost.model_breakdown': 'Model Dağılımı',
+  'cost.no_model_data': 'Model verisi yok.',
+  'cost.share': 'Pay',
+  'cost.load_failed': 'Maliyet verisi yüklenemedi',
+  'logs.title': 'Canlı Kayıtlar',
+  'logs.pause': 'Duraklat',
+  'logs.resume': 'Sürdür',
+  'logs.events': 'olay',
+  'logs.jump_to_bottom': 'Alta git',
+  'logs.filter_label': 'Filtre:',
+  'logs.paused_stream': 'Kayıt akışı duraklatıldı.',
+  'logs.waiting_for_events': 'Olaylar bekleniyor...',
+  'doctor.title': 'Sistem Teşhisleri',
+  'doctor.run': 'Teşhisleri Çalıştır',
+  'doctor.running': 'Teşhisler çalıştırılıyor...',
+  'doctor.running_short': 'Çalışıyor...',
+  'doctor.running_hint': 'Bu birkaç saniye sürebilir.',
+  'doctor.warn': 'Uyarı',
+  'doctor.issues_found': 'Sorunlar Bulundu',
+  'doctor.warnings': 'Uyarılar',
+  'doctor.all_clear': 'Temiz',
+  'doctor.instructions': 'ZeroClaw kurulumunuzu kontrol etmek için "Teşhisleri Çalıştır" düğmesine tıklayın.',
+  'auth.pair': 'Cihazı Eşle',
+  'auth.pair_button': 'Eşle',
+  'auth.logout': 'Çıkış Yap',
+  'auth.enter_code': 'Terminalinizdeki tek kullanımlık eşleştirme kodunu girin',
+  'auth.code_placeholder': '6 haneli kod',
+  'auth.pairing_progress': 'Eşleştiriliyor...',
+  'common.languages': 'Diller',
+  'common.select_language': 'Dil seçin',
+  'common.connecting': 'Bağlanıyor...',
+  'header.dashboard_tagline': 'ZeroClaw paneli',
+  'sidebar.gateway_dashboard': 'Ağ Geçidi + Panel',
+  'sidebar.runtime_mode': 'Çalışma Modu',
+  'navigation.open': 'Gezinmeyi aç',
+  'navigation.close': 'Gezinmeyi kapat',
+  'navigation.expand': 'Gezinmeyi genişlet',
+  'navigation.collapse': 'Gezinmeyi daralt',
+};
+
+export default tr;
