+++ web/src/lib/i18n/locales/uk.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const uk: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Панель',
+  'nav.agent': 'Агент',
+  'nav.tools': 'Інструменти',
+  'nav.cron': 'Заплановані завдання',
+  'nav.integrations': 'Інтеграції',
+  'nav.memory': 'Пам\u2019ять',
+  'nav.config': 'Конфігурація',
+  'nav.cost': 'Витрати',
+  'nav.logs': 'Журнали',
+  'nav.doctor': 'Діагностика',
+  'dashboard.hero_title': 'Електрична панель runtime',
+  'agent.placeholder': 'Введіть повідомлення…',
+  'tools.search': 'Пошук інструментів…',
+  'cron.add': 'Додати завдання',
+  'memory.add_memory': 'Додати пам\u2019ять',
+  'config.save': 'Зберегти',
+  'cost.token_statistics': 'Статистика токенів',
+  'logs.title': 'Живі журнали',
+  'doctor.title': 'Діагностика системи',
+  'auth.pair_button': 'З\u2019єднати',
+  'auth.enter_code': 'Введіть одноразовий код з\u2019єднання з термінала',
+  'auth.code_placeholder': '6-значний код',
+  'auth.pairing_progress': 'З\u2019єднання…',
+  'auth.logout': 'Вийти',
+  'common.languages': 'Мови',
+  'common.select_language': 'Оберіть мову',
+  'header.dashboard_tagline': 'Панель ZeroClaw',
+};
+
+export default uk;
