+++ web/src/lib/i18n/locales/ur.ts
@@ -1,0 +1,33 @@
+import type { TranslationKeys } from './en';
+
+const ur: Partial<TranslationKeys> = {
+  'nav.dashboard': 'ڈیش بورڈ',
+  'nav.agent': 'ایجنٹ',
+  'nav.tools': 'ٹولز',
+  'nav.cron': 'شیڈول شدہ کام',
+  'nav.integrations': 'انضمامات',
+  'nav.memory': 'میموری',
+  'nav.config': 'ترتیبات',
+  'nav.cost': 'لاگت',
+  'nav.logs': 'لاگز',
+  'nav.doctor': 'تشخیص',
+  'dashboard.hero_title': 'الیکٹرک رن ٹائم ڈیش بورڈ',
+  'agent.placeholder': 'پیغام لکھیں…',
+  'tools.search': 'ٹولز تلاش کریں…',
+  'cron.add': 'کام شامل کریں',
+  'memory.add_memory': 'میموری شامل کریں',
+  'config.save': 'محفوظ کریں',
+  'cost.token_statistics': 'ٹوکن کے اعدادوشمار',
+  'logs.title': 'لائیو لاگز',
+  'doctor.title': 'سسٹم تشخیص',
+  'auth.pair_button': 'جوڑیں',
+  'auth.enter_code': 'ٹرمینل سے ایک بار استعمال ہونے والا پیئرنگ کوڈ درج کریں',
+  'auth.code_placeholder': '6 ہندسوں کا کوڈ',
+  'auth.pairing_progress': 'جوڑا جا رہا ہے…',
+  'auth.logout': 'لاگ آؤٹ',
+  'common.languages': 'زبانیں',
+  'common.select_language': 'زبان منتخب کریں',
+  'header.dashboard_tagline': 'ZeroClaw ڈیش بورڈ',
+};
+
+export default ur;
