+++ web/src/lib/i18n/locales/vi.ts
@@ -1,0 +1,51 @@
+import type { TranslationKeys } from './en';
+
+const vi: Partial<TranslationKeys> = {
+  'nav.dashboard': 'Bảng điều khiển',
+  'nav.agent': 'Tác tử',
+  'nav.tools': 'Công cụ',
+  'nav.cron': 'Lịch tác vụ',
+  'nav.integrations': 'Tích hợp',
+  'nav.memory': 'Bộ nhớ',
+  'nav.config': 'Cấu hình',
+  'nav.cost': 'Chi phí',
+  'nav.logs': 'Nhật ký',
+  'nav.doctor': 'Chẩn đoán',
+  'dashboard.hero_title': 'Bảng điều khiển runtime điện xanh',
+  'dashboard.live_gateway': 'Cổng hoạt động',
+  'dashboard.unpaired': 'Chưa ghép đôi',
+  'agent.title': 'Trò chuyện với tác tử',
+  'agent.placeholder': 'Nhập tin nhắn…',
+  'agent.connecting': 'Đang kết nối…',
+  'agent.connected': 'Đã kết nối',
+  'agent.disconnected': 'Đã ngắt kết nối',
+  'tools.search': 'Tìm công cụ…',
+  'tools.agent_tools': 'Công cụ tác tử',
+  'tools.cli_tools': 'Công cụ CLI',
+  'cron.add': 'Thêm tác vụ',
+  'cron.scheduled_tasks': 'Tác vụ đã lên lịch',
+  'integrations.title': 'Tích hợp',
+  'memory.add_memory': 'Thêm bộ nhớ',
+  'memory.search_entries': 'Tìm trong bộ nhớ…',
+  'config.save': 'Lưu',
+  'config.saving': 'Đang lưu…',
+  'cost.session': 'Chi phí phiên',
+  'cost.daily': 'Chi phí ngày',
+  'cost.monthly': 'Chi phí tháng',
+  'logs.title': 'Nhật ký trực tiếp',
+  'logs.pause': 'Tạm dừng',
+  'logs.resume': 'Tiếp tục',
+  'doctor.title': 'Chẩn đoán hệ thống',
+  'doctor.run': 'Chạy chẩn đoán',
+  'doctor.running_short': 'Đang chạy…',
+  'auth.pair_button': 'Ghép đôi',
+  'auth.enter_code': 'Nhập mã ghép đôi một lần từ terminal',
+  'auth.code_placeholder': 'Mã 6 chữ số',
+  'auth.pairing_progress': 'Đang ghép đôi…',
+  'auth.logout': 'Đăng xuất',
+  'common.languages': 'Ngôn ngữ',
+  'common.select_language': 'Chọn ngôn ngữ',
+  'header.dashboard_tagline': 'Bảng điều khiển ZeroClaw',
+};
+
+export default vi;
