+++ web/src/lib/i18n/locales/zh-CN.ts
@@ -1,0 +1,51 @@
+import type { TranslationKeys } from './en';
+
+const zhCN: Partial<TranslationKeys> = {
+  'nav.dashboard': '仪表盘',
+  'nav.agent': '代理',
+  'nav.tools': '工具',
+  'nav.cron': '定时任务',
+  'nav.integrations': '集成',
+  'nav.memory': '记忆',
+  'nav.config': '配置',
+  'nav.cost': '成本跟踪',
+  'nav.logs': '日志',
+  'nav.doctor': '诊断',
+  'dashboard.hero_title': '电光运行仪表盘',
+  'dashboard.live_gateway': '在线网关',
+  'dashboard.unpaired': '未配对',
+  'agent.title': '代理聊天',
+  'agent.placeholder': '输入消息…',
+  'agent.connecting': '正在连接…',
+  'agent.connected': '已连接',
+  'agent.disconnected': '已断开',
+  'tools.search': '搜索工具…',
+  'tools.agent_tools': '代理工具',
+  'tools.cli_tools': 'CLI 工具',
+  'cron.add': '添加任务',
+  'cron.scheduled_tasks': '定时任务',
+  'integrations.title': '集成',
+  'memory.add_memory': '添加记忆',
+  'memory.search_entries': '搜索记忆条目…',
+  'config.save': '保存',
+  'config.saving': '正在保存…',
+  'cost.session': '会话成本',
+  'cost.daily': '每日成本',
+  'cost.monthly': '每月成本',
+  'logs.title': '实时日志',
+  'logs.pause': '暂停',
+  'logs.resume': '继续',
+  'doctor.title': '系统诊断',
+  'doctor.run': '运行诊断',
+  'doctor.running_short': '运行中…',
+  'auth.pair_button': '配对',
+  'auth.enter_code': '输入终端中的一次性配对码',
+  'auth.code_placeholder': '6 位代码',
+  'auth.pairing_progress': '正在配对…',
+  'auth.logout': '退出',
+  'common.languages': '语言',
+  'common.select_language': '选择语言',
+  'header.dashboard_tagline': 'ZeroClaw 仪表盘',
+};
+
+export default zhCN;
